# server/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PTCF Room & Equipment Reservation System API",
    "description": "API documentation for the UPLB ICropS Plant Tissue Culture Facility reservation management system.",
    "version": "1.0.0",
    "contact": {
      "name": "PTCF Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Local development server"
    },
    {
      "url": "https://your-render-url.onrender.com/api",
      "description": "Production server (Render)"
    }
  ],
...
  "paths": {
...
    "/auth/password-reset-request": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Request Supabase password reset email",
        "description": "Sends a Supabase Auth password reset email when AUTH_PROVIDER=supabase. The response is intentionally generic to avoid account enumeration.",
        "operationId": "requestPasswordReset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              },
              "examples": {
                "reset": {
                  "summary": "Request reset email",
                  "value": {
                    "email": "student@uplb.edu.ph",
                    "redirectTo": "http://localhost:5173/reset-password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "message": "If that email is registered, a password reset link will be sent."
                }
              }
            }
          },
          "400": {
            "description": "Missing email or Supabase Auth disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "email is required"
                }
              }
            }
          },
          "502": {
            "description": "Supabase Auth request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all users (Admin only)",
        "description": "Returns all registered users with their id, email, accountType, userCategory, and createdAt. Requires system_admin role.",
        "operationId": "listUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of user objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "email": {
                        "type": "string",
                        "example": "user@uplb.edu.ph"
                      },
                      "accountType": {
                        "type": "string",
                        "example": "regular_user"
                      },
                      "userCategory": {
                        "type": "string",
                        "example": "student"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
    "/auth/email-verification/resend": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Resend verification email",
        "description": "Resends a Supabase signup verification email when AUTH_PROVIDER=supabase. Response is intentionally generic.",
        "operationId": "resendEmailVerification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationResendRequest"
              },
              "examples": {
                "resend": {
                  "summary": "Resend verification link",
                  "value": {
                    "email": "student@uplb.edu.ph",
                    "redirectTo": "http://localhost:5173/login"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification resend request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "message": "If that email is pending verification, a new verification link will be sent."
                }
              }
            }
          },
          "400": {
            "description": "Missing email or Supabase disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "email is required"
                }
              }
            }
          },
          "502": {
            "description": "Supabase resend request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/oauth/start": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Start OAuth sign-in",
        "description": "Starts Supabase OAuth flow and returns provider redirect URL.",
        "operationId": "startOAuth",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthStartRequest"
              },
              "examples": {
                "google": {
                  "summary": "Start Google OAuth",
                  "value": {
                    "provider": "google",
                    "redirectTo": "http://localhost:5173/oauth/callback"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OAuth URL returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unsupported provider or Supabase disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "OAuth startup failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/oauth/exchange": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Complete OAuth sign-in",
        "description": "Exchanges Supabase OAuth tokens for normalized app session payload and local profile mapping.",
        "operationId": "exchangeOAuth",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthExchangeRequest"
              },
              "examples": {
                "exchange": {
                  "summary": "Exchange OAuth token",
                  "value": {
                    "accessToken": "supabase-access-token",
                    "refreshToken": "supabase-refresh-token"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OAuth exchange successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing token or Supabase disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid OAuth access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "OAuth exchange failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid credentials"
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
      "PasswordResetRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "student@uplb.edu.ph"
          },
          "redirectTo": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "Optional password reset redirect URL. Defaults to SUPABASE_PASSWORD_RESET_REDIRECT_URL, CLIENT_PASSWORD_RESET_URL, or CLIENT_URL.",
            "example": "http://localhost:5173/reset-password"
          }
        }
      },
...
      "EmailVerificationResendRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "student@uplb.edu.ph"
          },
          "redirectTo": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "example": "http://localhost:5173/login"
          }
        }
      },
      "OAuthStartRequest": {
        "type": "object",
        "required": [
          "provider"
        ],
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "google"
            ],
            "example": "google"
          },
          "redirectTo": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "example": "http://localhost:5173/oauth/callback"
          }
        }
      },
      "OAuthStartResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://<project-ref>.supabase.co/auth/v1/authorize?..."
          },
          "provider": {
            "type": "string",
            "example": "google"
          }
        }
      },
      "OAuthExchangeRequest": {
        "type": "object",
        "required": [
          "accessToken"
        ],
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "supabase-access-token"
          },
          "refreshToken": {
            "type": "string",
            "nullable": true,
            "example": "supabase-refresh-token"
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized - Missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "missingToken": {
                "value": {
                  "message": "Missing Bearer token"
                }
              },
              "invalidToken": {
                "value": {
                  "message": "Invalid or expired token"
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden - Insufficient role (requires staff or admin)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Forbidden: insufficient role"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter ONLY your JWT token (without 'Bearer' prefix). Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      }
    }
  },
  "security": []
}